+++ src/main/java/roomescape/auth/controller/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package roomescape.auth.web.controller.dto;
+package roomescape.auth.controller.dto;
 
 import jakarta.validation.constraints.NotNull;
