+++ src/main/java/roomescape/auth/web/resolver/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package roomescape.auth.annotation;
+package roomescape.auth.web.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
